requirements.txt
requests>=2.32.3
feedparser>=6.0.11
beautifulsoup4>=4.12.3
python-dateutil>=2.9.0.post0
pandas>=2.2.3
pyyaml>=6.0.2
